gaia store debugging
